+++ src/VariableManager.php
@@ -2,8 +2,6 @@
 
 namespace Fomvasss\Variable;
 
-use Illuminate\Cache\Repository as CacheRepository;
-
 /**
  * Class VariableManager
  *
@@ -13,33 +11,39 @@ class VariableManager implements VariableManagerContract
 {
     protected $variableModel;
 
-    protected $cacheRepo;
+    protected $cacheRepository;
 
-    protected $locale = null;
+    protected $locale;
 
-    private $cacheName;
-
-    protected $cacheTime;
+    protected $variables;
 
     /**
      * VariableManager constructor.
-     *
-     * @param \Illuminate\Cache\Repository $cacheRepo
+     * @param $cacheRepository
+     * @param null $app
      */
-    public function __construct(CacheRepository $cacheRepo)
+    public function __construct($variableModel, $cacheRepository, $app = null)
     {
-        $this->cacheTime = config('variables.cache.time', 360);
-        $this->cacheName = config('variables.cache.name', 'laravel.variables.cache');
-        $this->variableModel = $this->model();
-        $this->cacheRepo = $cacheRepo;
+        if (! $app) {
+            $app = app();   //Fallback when $app is not given
+        }
+        $this->app = $app;
+
+        $this->config = $this->app['config']->get('variables');
+
+        $this->variableModel = $variableModel;
+
+        $this->cacheRepository = $cacheRepository;
     }
 
     /**
-     * @return mixed
+     * @param string $key
+     * @param string|null $default
+     * @return mixed|string
      */
-    protected function model()
+    public function get(string $key, $default = null)
     {
-        return app()->make(config('variables.model', \Fomvasss\Variable\Variable::class));
+        return $this->all()[$key] ?? $default;
     }
 
     /**
@@ -47,69 +51,22 @@ class VariableManager implements VariableManagerContract
      */
     public function all(): array
     {
-        $settings = $this->getAll();
-        return $settings->pluck('value', 'name')->toArray();
-    }
+        if (! isset($this->variables)) {
 
-    /**
-     * @param $name
-     * @param null $default
-     * @return null
-     */
-    public function first($name, $default = null)
-    {
-        if (!empty($variable = $this->firstByName($name))) {
-            return $variable->value;
+            $all = $this->cacheRepository->remember($this->config['cache']['name'], $this->config['cache']['time'], function () {
+                return $this->variableModel->select('key', 'value', 'locale')->get();
+            });
+
+            $allByLocale = $this->locale ? $all->where('locale', $this->locale) : $all;
+
+            $this->variables = $allByLocale->pluck('value', 'key')->toArray();
         }
 
-        return $default;
+        return $this->variables;
     }
 
     /**
-     * @param $name
-     * @param bool $asoc
-     * @return array|mixed
-     */
-    public function firstJsonDecode($name, $asoc = true)
-    {
-        if (!empty($variable = $this->firstByName($name))) {
-            return json_decode($variable->value, $asoc);
-        }
-
-        return [];
-    }
-
-    /**
-     * @param $name
-     * @param null $value
-     * @param null $description
-     * @return int
-     */
-    public function set($name, $value = null, $description = null): int
-    {
-        $this->cacheRepo->forget($this->cacheName);
-
-        return $this->variableModel->updateOrCreate([
-            'name' => $name, 'locale' => $this->locale
-        ], [
-            'value' => $value, 'description' => $description,
-        ]) ? 1 : 0;
-    }
-
-    /**
-     * @param $name
-     * @return int
-     */
-    public function delete($name): int
-    {
-        $res = $this->variableModel->where('name', $name)->where('locale', $this->locale)->delete();
-        $this->cacheRepo->forget($this->cacheName);
-
-        return $res ? 0 : 1;
-    }
-
-    /**
-     * @param string|null $locale
+     * @param string $locale
      * @return $this
      */
     public function locale(string $locale = null)
@@ -117,45 +74,5 @@ class VariableManager implements VariableManagerContract
         $this->locale = $locale;
 
         return $this;
-    }
-
-    /**
-     * @param array $attributes
-     * @return int
-     */
-    public function setArray(array $attributes): int
-    {
-        $r = 0;
-        foreach ($attributes as $name => $value) {
-            $this->variableModel->updateOrCreate([
-                'name' => $name, 'locale' => $this->locale
-            ], [
-                'value' => $value
-            ]);
-            $r++;
-        }
-        $this->cacheRepo->forget($this->cacheName);
-        return $r;
-    }
-
-    /**
-     * @param $name
-     * @return mixed
-     */
-    protected function firstByName($name)
-    {
-        return $this->getAll()->where('name', $name)->where('locale', $this->locale)->first();
-    }
-
-    /**
-     * @return mixed
-     */
-    protected function getAll()
-    {
-        $settings = $this->cacheRepo->remember($this->cacheName, $this->cacheTime, function () {
-            return $this->variableModel->all();
-        });
-
-        return $this->locale ? $settings->where('locale', $this->locale) : $settings;
     }
 }